refactor: ImageExtension을 사용하는 Service 폴더로 이동
